Add support for Areas in core renderers
