+++ packages/core/src/renderers/area/index.ts
@@ -1,5 +1,7 @@
+import { IStackedAreaFocusProps, stackedFocus } from "./stackedFocus";
+import { IStackedAreaTooltipProps, stackedTooltip } from "./stackedTooltip";
 import { canvas as canvasRender } from "./canvas";
-import { stackedCanvas as stackedCanvasRender } from "./stackedCanvas";
+import { stackedCanvas } from "./stackedCanvas";
 import { stacked as stackedRender } from "./stacked";
 
 export const canvas = {
@@ -9,9 +11,14 @@ export const canvas = {
 export const stacked = {
     render: stackedRender,
     canvas: {
-        render: stackedCanvasRender,
+        render: stackedCanvas,
     },
+    focus: stackedFocus,
+    tooltip: stackedTooltip,
 };
 
 export * from "./render";
 export * from "./focus";
+export * from "./tooltip";
+export type { IStackedAreaFocusProps };
+export type { IStackedAreaTooltipProps };
