+++ packages/types/src/IInvertScale.ts
@@ -1,0 +1,8 @@
+import type { IValue } from "./IValue";
+
+/**
+ * A scale with an invert function
+ */
+export interface IInvertScale {
+    invert?: (value: IValue) => number;
+}
