+++ packages/types/src/index.ts
@@ -24,3 +24,4 @@ export * from "./ITheme";
 export * from "./ITooltipItem";
 export * from "./IValue";
 export * from "./IFocused";
+export * from "./IInvertScale";
